requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
scikit-image
python-multipart
